Fixed minor test issue
I was having some trouble testing some of the plugins, due to odd
inconsistent results. I realized the cause of this was that by default,
gradle assumes that tests will pass if the test files haven't changed
since the last time they passed. As this is pretty much never the case,
I disabled this behavior.