Fixed a couple of things in the buildscript
Added a repositories/maven block in the root (gradle was complaining
that it wasn't there).

Added a feature that makes a softlink artifactory/artifactory to the
artifactory instance in artifactory/artifactory-<version>, so that you
can use version-independent paths into the instance.

Fixed stopWorkOnPlugin: If the user passes 'all' and there are no
plugins loaded, this is not an error.

Fixed stopWorkOnPlugin: the plugin reload doesn't actually do anything,
and sometimes causes the job to fail for no good reason. This has been
removed.